Update csv2sqlite.py
Added reminder to help statement for file path entry: use forward slashes, not backslashes.

In my first use of this program (as a relative newbie to Python) this was non-obvious, even with the error messages I was receiving. I realise there are other options, but this is the easiest to fit in the space.